+++ etc/gtfs/postprocessing.d/41-postgrest-domain-text-plain.sql
@@ -1,0 +1,2 @@
+-- see https://postgrest.org/en/stable/references/api/media_type_handlers.html
+CREATE DOMAIN "text/plain" AS text;