+++ etc/gtfs/postprocessing.d/42-postgrest-imported-at-metric.sql
@@ -1,0 +1,5 @@
+CREATE OR REPLACE FUNCTION api.gtfs_data_imported_at_metric ()
+RETURNS "text/plain"
+AS $$
+	SELECT 'gtfs_data_imported_at=' || floor(EXTRACT(EPOCH FROM api.gtfs_data_imported_at())) || E'\n';
+$$ LANGUAGE SQL IMMUTABLE;