+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
+## [Unreleased]
+
+### Added
+
+- `gtfs-api`: Expose GTFS import date+time (`SELECT gtfs_data_imported_at()`) as a Prometheus metric under `/gtfs/rpc/gtfs_data_imported_at_metric`.
+
 ## 2024-12-04
 
 ### Changed